change header location string to not contain capital letter L
The password_verify didn't seem to work at first, but after some research I discovered that the database varchar length was set to 30 characters but the bcrypt password hash was longer which is why the password wouldn't get accepted. I changed this in the database and tested it. The login feature now works just fine.