add editPassword page which allows the user to change their password by following the check up procedure from register
The user has to fill in their current password which gets compared to the hash in the database. Then the newPassword and newPasswordConfirmation get checked to be equal and to see if the password is long enough. When the button is pressed, the data is sent to the database and the user can now use the new password