add editProfile page were user can change their info
The current data gets loaded into the input fields as data so the user can make small changes if needed, and all data is always filled in in case the user forgets to change something. The email address is checked to be a valid email address and has to be filled in. This data is also updated based on the session email, which gets updated to the email address from the input field after pressing the edit profile button