add a check to see which like icon image needs to be loaded + add ajax like and icon update
When the page is loaded, all currently liked posts need to have a coloured like icon to show that that page has already been liked. When a user presses a button, the current like number needs to be increased or decreased based on the liked_status and the icon needs to change colour.