remove vardump from likePost.php and change res.data to res.data['status'] to get the correct response
The vardump got in the way of the response and added an array to the response item. By removing this, I was able to get the correct response to check what action has to be taken on the like button and like counter. All that is left is to get the correct js way to change the current icon and like counter